Delegate groupfigures and ungroupfigures
