+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/action/grouping/Grouping.java
@@ -1,0 +1,41 @@
+package org.jhotdraw.draw.action.grouping;
+
+import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
+import org.jhotdraw.draw.DrawingView;
+import org.jhotdraw.draw.figure.CompositeFigure;
+import org.jhotdraw.draw.figure.Figure;
+
+import java.util.Collection;
+import java.util.LinkedList;
+
+public class Grouping {
+    public Grouping() {
+    }
+
+    @FeatureEntryPoint(value = "Group figures")
+    public void groupFigures(DrawingView view, CompositeFigure group, Collection<Figure> figures) {
+        Collection<Figure> sorted = view.getDrawing().sort(figures);
+        int index = view.getDrawing().indexOf(sorted.iterator().next());
+        view.getDrawing().basicRemoveAll(figures);
+        view.clearSelection();
+        view.getDrawing().add(index, group);
+        group.willChange();
+        for (Figure f : sorted) {
+            f.willChange();
+            group.basicAdd(f);
+        }
+        group.changed();
+        view.addToSelection(group);
+    }
+
+    @FeatureEntryPoint(value = "Ungroup figures")
+    public Collection<Figure> ungroupFigures(DrawingView view, CompositeFigure group) {
+        LinkedList<Figure> figures = new LinkedList<>(group.getChildren());
+        view.clearSelection();
+        group.basicRemoveAllChildren();
+        view.getDrawing().basicAddAll(view.getDrawing().indexOf(group), figures);
+        view.getDrawing().remove(group);
+        view.addToSelection(figures);
+        return figures;
+    }
+}
