+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/action/grouping/strategy/GroupStrategy.java
@@ -2,6 +2,7 @@ package org.jhotdraw.draw.action.grouping.strategy;
 
 import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
 import org.jhotdraw.draw.DrawingView;
+import org.jhotdraw.draw.action.grouping.Grouping;
 import org.jhotdraw.draw.figure.CompositeFigure;
 import org.jhotdraw.draw.figure.Figure;
 
@@ -10,6 +11,7 @@ import java.util.LinkedList;
 
 public class GroupStrategy implements GroupingStrategy {
     public static final String ID = "edit.groupSelection";
+    private final Grouping grouping = new Grouping();
 
     @Override
     public String getID() {
@@ -26,23 +28,6 @@ public class GroupStrategy implements GroupingStrategy {
         final LinkedList<Figure> ungroupedFigures = new LinkedList<>(figures);
         final CompositeFigure group = (CompositeFigure) prototype.clone();
 
-        groupFigures(drawingView, group, ungroupedFigures);
-    }
-
-
-    @FeatureEntryPoint(value = "Group figures")
-    public void groupFigures(DrawingView view, CompositeFigure group, Collection<Figure> figures) {
-        Collection<Figure> sorted = view.getDrawing().sort(figures);
-        int index = view.getDrawing().indexOf(sorted.iterator().next());
-        view.getDrawing().basicRemoveAll(figures);
-        view.clearSelection();
-        view.getDrawing().add(index, group);
-        group.willChange();
-        for (Figure f : sorted) {
-            f.willChange();
-            group.basicAdd(f);
-        }
-        group.changed();
-        view.addToSelection(group);
+        grouping.groupFigures(drawingView, group, ungroupedFigures);
     }
 }
